Port to Fluid and QtMultimedia
Port the UI to QtQuick Controls 2.0 and Fluid, remove Vlc dependency
in favor of QtMultimedia that shoud work on various platforms.

Convert the build system to CMake and refactor the source tree, just like
any other Liri project

The user interface has also changed a bit: gone is the settings drawer,
the only settings you could interact with were the primary and accent
colors but that is something that we want to set by design.

Closes: #3
Closes: #4
Closes: #5